Correct some numeric type and guard objects
- Avoid casting as much as possible
- Change some numeric types to match its expected ranges
- Utilize static_cast explicitly
- Encapsulate newly instantiated objects inside `std::unique_ptr`

+++ modules/Core/include/Open3D/Core/Geometry/LineSet.h
@@ -39,7 +39,7 @@ class PointCloud;
 class LineSet : public Geometry3D
 {
 public:
-    typedef std::pair<int, int> LineSegment;
+    typedef std::pair<size_t, size_t> LineSegment;
 
 public:
     LineSet() : Geometry3D(Geometry::GeometryType::LineSet) {}
@@ -85,6 +85,6 @@ public:
 /// correspondence set (LineSetFactory.cpp)
 std::shared_ptr<LineSet> CreateLineSetFromPointCloudCorrespondences(
         const PointCloud &cloud0, const PointCloud &cloud1,
-        const std::vector<std::pair<int, int>> &correspondences);
+        const std::vector<std::pair<size_t, size_t>> &correspondences);
 
 }   // namespace open3d
