+++ modules/Core/include/Open3D/Core/Integration/UniformTSDFVolume.h
@@ -33,7 +33,7 @@ namespace open3d {
 
 class UniformTSDFVolume : public TSDFVolume {
 public:
-    UniformTSDFVolume(double length, int resolution, double sdf_trunc,
+    UniformTSDFVolume(double length, uint32_t resolution, double sdf_trunc,
             bool with_color, const Eigen::Vector3d &origin = Eigen::Vector3d::Zero());
     ~UniformTSDFVolume() override;
 
@@ -55,19 +55,19 @@ public:
             const Eigen::Matrix4d &extrinsic,
             const Image &depth_to_camera_distance_multiplier);
 
-    inline int IndexOf(int x, int y, int z) const {
+    inline size_t IndexOf(int32_t x, int32_t y, int32_t z) const {
         return x * resolution_ * resolution_ + y * resolution_ + z;
     }
 
-    inline int IndexOf(const Eigen::Vector3i &xyz) const {
+    inline size_t IndexOf(const Eigen::Vector3i &xyz) const {
         return IndexOf(xyz(0), xyz(1), xyz(2));
     }
 
 public:
     Eigen::Vector3d origin_;
     double length_;
-    int resolution_;
-    int voxel_num_;
+    uint32_t resolution_;
+    uint32_t voxel_num_;
     std::vector<float> tsdf_;
     std::vector<Eigen::Vector3f> color_;
     std::vector<float> weight_;
