+++ modules/Core/include/Open3D/Core/Registration/FastGlobalRegistration.h
@@ -46,9 +46,9 @@ public:
             bool use_absolute_scale = false,
             bool decrease_mu = true,
             double maximum_correspondence_distance = 0.025,
-            int iteration_number = 64,
+            size_t iteration_number = 64,
             double tuple_scale = 0.95,
-            int maximum_tuple_count = 1000) :
+            uint32_t maximum_tuple_count = 1000) :
             division_factor_(division_factor),
             use_absolute_scale_(use_absolute_scale),
             iteration_number_(iteration_number),
@@ -65,11 +65,11 @@ public:
     // Maximum correspondence distance (also see comment of USE_ABSOLUTE_SCALE)
     double maximum_correspondence_distance_;
     // Maximum number of iteration
-    int iteration_number_;
+    size_t iteration_number_;
     // Similarity measure used for tuples of feature points.
     double tuple_scale_;
     // Maximum tuple numbers.
-    int maximum_tuple_count_;
+    size_t maximum_tuple_count_;
 };
 
 RegistrationResult FastGlobalRegistration(
