+++ modules/Core/src/Geometry/ImageFactory.cpp
@@ -37,25 +37,25 @@ std::shared_ptr<Image> CreateDepthToCameraDistanceMultiplierFloatImage(
     auto fimage = std::make_shared<Image>();
     fimage->PrepareImage(intrinsic.width_, intrinsic.height_, 1, 4);
     float ffl_inv[2] = {
-            1.0f / (float)intrinsic.GetFocalLength().first,
-            1.0f / (float)intrinsic.GetFocalLength().second,
+            1.0f / static_cast<float>(intrinsic.GetFocalLength().first),
+            1.0f / static_cast<float>(intrinsic.GetFocalLength().second),
     };
     float fpp[2] = {
-            (float)intrinsic.GetPrincipalPoint().first,
-            (float)intrinsic.GetPrincipalPoint().second,
+            static_cast<float>(intrinsic.GetPrincipalPoint().first),
+            static_cast<float>(intrinsic.GetPrincipalPoint().second),
     };
     std::vector<float> xx(intrinsic.width_);
     std::vector<float> yy(intrinsic.height_);
-    for (int j = 0; j < intrinsic.width_; j++) {
+    for (int32_t j = 0; j < intrinsic.width_; j++) {
         xx[j] = (j - fpp[0]) * ffl_inv[0];
     }
-    for (int i = 0; i < intrinsic.height_; i++) {
+    for (int32_t i = 0; i < intrinsic.height_; i++) {
         yy[i] = (i - fpp[1]) * ffl_inv[1];
     }
-    for (int i = 0; i < intrinsic.height_; i++) {
+    for (int32_t i = 0; i < intrinsic.height_; i++) {
         float *fp = (float *)(fimage->data_.data() +
                 i * fimage->BytesPerLine());
-        for (int j = 0; j < intrinsic.width_; j++, fp++) {
+        for (int32_t j = 0; j < intrinsic.width_; j++, fp++) {
             *fp = sqrtf(xx[j] * xx[j] + yy[i] * yy[i] + 1.0f);
         }
     }
@@ -70,17 +70,17 @@ std::shared_ptr<Image> CreateFloatImageFromImage(const Image &image,
         return fimage;
     }
     fimage->PrepareImage(image.width_, image.height_, 1, 4);
-    for (int i = 0; i < image.height_ * image.width_; i++) {
+    for (int32_t i = 0; i < image.height_ * image.width_; i++) {
         float *p = (float *)(fimage->data_.data() + i * 4);
         const uint8_t *pi = image.data_.data() +
                 i * image.num_of_channels_ * image.bytes_per_channel_;
         if (image.num_of_channels_ == 1) {
             // grayscale image
             if (image.bytes_per_channel_ == 1) {
-                *p = (float)(*pi) / 255.0f;
+                *p = static_cast<float>(*pi) / 255.0f;
             } else if (image.bytes_per_channel_ == 2) {
                 const uint16_t *pi16 = (const uint16_t *)pi;
-                *p = (float)(*pi16);
+                *p = static_cast<float>(*pi16);
             } else if (image.bytes_per_channel_ == 4) {
                 const float *pf = (const float *)pi;
                 *p = *pf;
@@ -88,23 +88,23 @@ std::shared_ptr<Image> CreateFloatImageFromImage(const Image &image,
         } else if (image.num_of_channels_ == 3) {
             if (image.bytes_per_channel_ == 1) {
                 if (type == Image::ColorToIntensityConversionType::Equal) {
-                    *p = ((float)(pi[0]) + (float)(pi[1]) + (float)(pi[2])) /
+                    *p = (static_cast<float>(pi[0]) + static_cast<float>(pi[1]) + static_cast<float>(pi[2])) /
                         3.0f / 255.0f;
                 } else if (type ==
                         Image::ColorToIntensityConversionType::Weighted) {
-                    *p = (0.2990f * (float)(pi[0]) + 0.5870f * (float)(pi[1]) +
-                            0.1140f * (float)(pi[2])) / 255.0f;
+                    *p = (0.2990f * static_cast<float>(pi[0]) + 0.5870f * static_cast<float>(pi[1]) +
+                            0.1140f * static_cast<float>(pi[2])) / 255.0f;
                 }
             } else if (image.bytes_per_channel_ == 2) {
                 const uint16_t *pi16 = (const uint16_t *)pi;
                 if (type == Image::ColorToIntensityConversionType::Equal) {
-                    *p = ((float)(pi16[0]) + (float)(pi16[1]) +
-                            (float)(pi16[2])) / 3.0f;
+                    *p = (static_cast<float>(pi16[0]) + static_cast<float>(pi16[1]) +
+                            static_cast<float>(pi16[2])) / 3.0f;
                 } else if (type ==
                         Image::ColorToIntensityConversionType::Weighted) {
-                    *p = (0.2990f * (float)(pi16[0]) +
-                            0.5870f * (float)(pi16[1]) +
-                            0.1140f * (float)(pi16[2]));
+                    *p = (0.2990f * static_cast<float>(pi16[0]) +
+                            0.5870f * static_cast<float>(pi16[1]) +
+                            0.1140f * static_cast<float>(pi16[2]));
                 }
             } else if (image.bytes_per_channel_ == 4) {
                 const float *pf = (const float *)pi;
@@ -134,7 +134,7 @@ std::shared_ptr<Image> CreateImageFromFloatImage(const Image &input)
             input.width_, input.height_, input.num_of_channels_, sizeof(T));
     const float *pi = (const float *)input.data_.data();
     T *p = (T*)output->data_.data();
-    for (int i = 0; i < input.height_ * input.width_; i++, p++, pi++) {
+    for (int32_t i = 0; i < input.height_ * input.width_; i++, p++, pi++) {
         if (sizeof(T) == 1)
             *p = static_cast<T>(*pi * 255.0f);
         if (sizeof(T) == 2)
@@ -149,7 +149,7 @@ template std::shared_ptr<Image> CreateImageFromFloatImage<uint16_t>(
         const Image &input);
 
 ImagePyramid CreateImagePyramid(
-        const Image &input, size_t num_of_levels,
+        const Image &input, uint8_t num_of_levels,
         bool with_gaussian_filter /*= true*/)
 {
     std::vector<std::shared_ptr<Image>> pyramid_image;
@@ -159,7 +159,7 @@ ImagePyramid CreateImagePyramid(
         return pyramid_image;
     }
 
-    for (int i = 0; i < num_of_levels; i++) {
+    for (uint32_t i = 0; i < num_of_levels; i++) {
         if (i == 0) {
             std::shared_ptr<Image> input_copy_ptr = std::make_shared<Image>();
             *input_copy_ptr = input;
