+++ modules/Core/src/Registration/TransformationEstimation.cpp
@@ -81,7 +81,7 @@ Eigen::Matrix4d TransformationEstimationPointToPlane::ComputeTransformation(
         return Eigen::Matrix4d::Identity();
 
     auto compute_jacobian_and_residual = [&]
-            (int i, Eigen::Vector6d &J_r, double &r) {
+            (size_t i, Eigen::Vector6d &J_r, double &r) {
         const Eigen::Vector3d &vs = source.points_[corres[i][0]];
         const Eigen::Vector3d &vt = target.points_[corres[i][1]];
         const Eigen::Vector3d &nt = target.normals_[corres[i][1]];
@@ -93,7 +93,7 @@ Eigen::Matrix4d TransformationEstimationPointToPlane::ComputeTransformation(
     Eigen::Matrix6d JTJ;
     Eigen::Vector6d JTr;
     std::tie(JTJ, JTr) = ComputeJTJandJTr<Eigen::Matrix6d, Eigen::Vector6d>(
-            compute_jacobian_and_residual, (int)corres.size());
+            compute_jacobian_and_residual, corres.size());
 
     bool is_success;
     Eigen::Matrix4d extrinsic;
